refactor(cli): route processors through pipelines entrypoints
